refact: 화면 조회 수정
