+++ src/Components/DiaryCard/DiaryCard.module.scss
@@ -1,22 +1,3 @@
-.all-list {
-  display: flex;
-  flex-direction: column;
-  gap: 4.5rem;
-}
-
-.day-list-section {
-  display: flex;
-  flex-direction: column;
-  gap: 1.5rem;
-}
-
-.diary-list {
-  display: flex;
-  flex-direction: column;
-  gap: 1.5rem;
-  user-select: none;
-}
-
 .diary-item {
   width: 100%;
   border-radius: 0.5rem;
