+++ src/Components/DiaryCard/index.ts
@@ -1,0 +1,1 @@
+export { DiaryCard } from './DiaryCard';
