+++ src/Layouts/TabLayout/TabLayout.module.scss
@@ -2,14 +2,14 @@
   display: flex;
   flex-direction: column;
   flex: 1;
+  max-height: 100dvh;
   // width: 100%;
   // height: 100%;
 }
 
 .header {
-  display: flex;
-  justify-content: center;
-  align-items: center;
+  background-color: var(--white);
+  z-index: 1;
 }
 
 .side-layout {
@@ -17,6 +17,7 @@
   flex-direction: row;
   flex: 1;
   min-height: 0;
+  background-color: var(--gray100);
 }
 
 .tab-container {
@@ -96,4 +97,52 @@
     transform: translateX(0);
     opacity: 1;
   }
+}
+
+.container {
+  width: 100%;
+
+  display: flex;
+  flex-direction: column;
+
+  padding: 0.5rem 0 0 0.5rem;
+
+  background-color: var(--white);
+  box-shadow: 0 0 12px rgba(0, 0, 0, 0.1);
+  border-radius: 0.5rem 0 0 0.5rem;
+  overflow: hidden;
+}
+
+.dashed-container {
+  display: flex;
+  flex-direction: column;
+  gap: 3rem;
+
+  overflow-y: auto;
+  overflow-x: hidden;
+
+  width: 100%;
+  flex: 1;
+
+  border-top: 1px dashed var(--gray400);
+  border-left: 1px dashed var(--gray400);
+  border-radius: var(--size-1);
+
+  padding: 1rem 1rem 4.5rem;
+
+  scrollbar-width: none;
+  // &::-webkit-scrollbar {
+  //   width: 16px;
+  // }
+
+  // &::-webkit-scrollbar-track {
+  //   background-color: transparent;
+  // }
+
+  // &::-webkit-scrollbar-thumb {
+  //   border: 4px solid transparent;
+  //   background: #606060;
+  //   background-clip: content-box;
+  //   border-radius: 8px;
+  // }
 }
