+++ src/Pages/DiaryAllPage/DiaryAllPage.module copy.scss
@@ -1,0 +1,53 @@
+.container {
+  width: 100%;
+  height: 100%;
+  overflow-y: scroll;
+  margin: var(--size-1) 0 0 var(--size-1);
+  border-top: 1px dashed var(--gray400);
+  border-left: 1px dashed var(--gray400);
+  border-radius: var(--size-1);
+  &::-webkit-scrollbar {
+    width: 16px;
+  }
+
+  &::-webkit-scrollbar-track {
+    background-color: transparent;
+  }
+
+  &::-webkit-scrollbar-thumb {
+    border: 4px solid transparent;
+    background: #606060;
+    background-clip: content-box;
+    border-radius: 8px;
+  }
+  .header {
+    padding: 0 var(--size-5);
+    margin-top: var(--size-5);
+    h1 {
+      font-weight: 600;
+      font-size: 1.4375rem;
+      margin-bottom: 0.375rem;
+    }
+    p {
+      color: var(--gray600);
+      font-size: var(--size-3);
+      margin-bottom: var(--size-8);
+      font-weight: 400;
+    }
+  }
+  .content {
+    padding: 0 var(--size-5);
+    margin-top: var(--size-5);
+    div {
+      h3 {
+        font-weight: 500;
+        font-size: var(--size-5);
+        margin-bottom: var(--size-4);
+      }
+    }
+  }
+}
+
+.not-diary-container {
+  margin-top: 7rem;
+}
