+++ src/Pages/DiaryCalendarPage/DiaryCalendarPage.module.scss
@@ -1,124 +1,10 @@
-.container {
-  width: 100%;
-
-  display: flex;
-  flex-direction: column;
-
-  padding: 0.5rem 0 0 0.5rem;
-}
-
-.dashed-container {
-  display: flex;
-  flex-direction: column;
-  gap: 3rem;
-
-  overflow-y: auto;
-  overflow-x: hidden;
-
-  width: 100%;
-  flex: 1;
-
-  border-top: 1px dashed var(--gray400);
-  border-left: 1px dashed var(--gray400);
-  border-radius: var(--size-1);
-
-  padding: 1rem 1rem 4.5rem;
-
-  &::-webkit-scrollbar {
-    width: 16px;
-  }
-
-  &::-webkit-scrollbar-track {
-    background-color: transparent;
-  }
-
-  &::-webkit-scrollbar-thumb {
-    border: 4px solid transparent;
-    background: #606060;
-    background-clip: content-box;
-    border-radius: 8px;
-  }
-}
-
-.calendar-section {
-  display: flex;
-  flex-direction: column;
-  gap: 1rem;
-}
-
-.date-container {
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  align-items: center;
-
-  p {
-    font-size: 1.125rem;
-  }
-}
-
-.calendar-container {
-  display: grid;
-  grid-template-columns: repeat(7, 1fr);
-  row-gap: 0.75rem;
-  place-items: stretch;
-}
-
-.day-item {
-  list-style: none;
-  aspect-ratio: 1 / 1;
-
-  display: flex;
-  justify-content: center;
-  align-items: center;
-
-  user-select: none;
-  cursor: pointer;
-
-  position: relative;
-
-  &::before {
-    content: '';
-    position: absolute;
-    background-color: var(--primary);
-
-    width: 4px;
-    height: 4px;
-
-    bottom: 0;
-
-    border-radius: 9999px;
-
-    display: none;
-  }
-}
-
-.header-item {
-  font-weight: 600;
-}
-
-.has-item {
-  &::before {
-    display: block;
-  }
-}
-
-.active-item {
-  background-color: var(--primary);
-  border-radius: 9999px;
-  color: var(--white);
-}
-
-.grayed {
-  color: var(--gray400);
-}
-
 .no-diary-view {
   display: flex;
   flex-direction: column;
   justify-content: center;
   align-items: center;
   gap: 0.5rem;
+  padding-top: 3rem;
 }
 
 .size120 {
