+++ src/Pages/DiaryListPage/DiaryListPage.module copy.scss
@@ -1,0 +1,299 @@
+@mixin comment-reply-box {
+  display: flex;
+  img {
+    width: var(--size-10);
+    height: var(--size-10);
+    border-radius: 50%;
+    object-fit: cover;
+  }
+  .body {
+    div {
+      margin-left: 1.125rem;
+      color: var(--gray1000);
+      span {
+      }
+      p {
+        margin-top: var(--size-2);
+      }
+    }
+    .etc {
+      margin-top: var(--size-2);
+      color: var(--gray500);
+      font-size: 0.875rem;
+      button {
+        color: var(--gray500);
+        font-size: 0.875rem;
+      }
+    }
+  }
+  @content;
+}
+
+.container {
+  width: 100%;
+  height: 100%;
+  .header {
+    width: 100%;
+    height: var(--size-12);
+    font-weight: 600;
+    display: inline-flex;
+    button {
+      height: 100%;
+      width: 20%;
+      display: flex;
+      justify-content: center;
+      align-items: center;
+    }
+    div {
+      display: flex;
+      justify-content: center;
+      align-items: center;
+      height: 100%;
+      width: 60%;
+    }
+  }
+  .tablist-container {
+    width: 100%;
+    height: 3rem;
+    margin: var(--size-5) 0 var(--size-8) 0;
+    div {
+      width: calc(100% - var(--size-4));
+      white-space: nowrap;
+      overflow-x: auto;
+      margin: 0 auto;
+      &::-webkit-scrollbar {
+        width: 16px;
+      }
+
+      &::-webkit-scrollbar-track {
+        background-color: transparent;
+      }
+
+      &::-webkit-scrollbar-thumb {
+        border: 4px solid transparent;
+        background: #606060;
+        background-clip: content-box;
+        border-radius: 8px;
+      }
+      .member-tab {
+        display: flex;
+        align-items: end;
+        .background {
+          background: linear-gradient(180deg, rgba(0, 0, 0, 0) 0%, rgba(0, 0, 0, 0.04) 100%);
+          width: 100%;
+          height: var(--size-2);
+        }
+        .member {
+          margin-right: var(--size-1);
+          background: #fafafa;
+          box-shadow: 0.125rem 0 0.125rem rgba(0, 0, 0, 0.24);
+          border-radius: var(--size-2) var(--size-2) 0 0;
+          width: auto;
+          padding: 0.5313rem var(--size-5);
+          height: var(--size-9);
+        }
+
+        .member.active {
+          margin-right: var(--size-1);
+          color: #ffffff;
+          box-shadow: 0.125rem 0 0.125rem rgba(0, 0, 0, 0.24);
+          border-radius: var(--size-2) var(--size-2) 0 0;
+          width: auto;
+          padding: 0.5313rem var(--size-5);
+          height: 2.9rem;
+        }
+      }
+    }
+  }
+
+  .scroll-container {
+    width: 100%;
+    height: calc(100% - 6.25rem);
+    .content {
+      width: calc(100% - var(--size-5) * 2);
+      margin: 0 auto;
+      h2 {
+        font-weight: 500;
+        font-size: 1.4375rem;
+      }
+      .etc {
+        span {
+          color: var(--gray500);
+          font-weight: 400;
+          font-size: var(--size-3);
+        }
+        margin-bottom: var(--size-6);
+      }
+      .contents {
+        width: 100%;
+        color: var(--gray900);
+        font-size: 1.375rem;
+        font-weight: 300;
+        margin: var(--size-4) 0 var(--size-3) 0;
+        line-height: 2rem;
+        text-indent: 2.3rem;
+
+        .icons {
+          position: absolute;
+          top: 13rem;
+          left: calc(50% - 11.2rem);
+          float: left;
+          z-index: 1;
+        }
+        .img-box-container {
+          width: 100%;
+          white-space: nowrap;
+          overflow-x: hidden;
+          overflow-y: hidden;
+          margin-bottom: var(--size-4);
+          border: 1px solid var(--gray300);
+          .img-box {
+            width: 100%;
+            height: 17.5rem;
+            img {
+              width: 100%;
+              height: 100%;
+              object-fit: cover;
+              transform: translateX(-2.25rem);
+            }
+          }
+        }
+      }
+      .button {
+        color: var(--gray500);
+        font-size: 0.7rem;
+        margin: var(--size-2) 0;
+      }
+      .comment-count {
+        width: calc(100% + var(--size-5));
+        transform: translateX(-1.25rem);
+        font-size: var(--size-3);
+        color: var(--gray500);
+        margin-top: var(--size-11);
+        padding-bottom: var(--size-4);
+        border-bottom: 1px solid var(--gray300);
+        p {
+          margin-left: 1.25rem;
+          font-size: 0.7rem;
+        }
+      }
+    }
+  }
+
+  // scroll-container 내부의 comment-container
+  .comment-container {
+    width: 100%;
+    margin: var(--size-6) 0 2rem 0;
+    padding-bottom: 3rem;
+    height: auto;
+    .comment-area {
+      .comment-box {
+        margin: var(--size-2) 0 var(--size-2) var(--size-5);
+        @include comment-reply-box();
+      }
+      .reply-box {
+        margin: var(--size-2) 0 var(--size-2) 4.5rem;
+        @include comment-reply-box();
+      }
+    }
+  }
+  // scroll-container 내부의 comment-container 내부의 write-container
+  .write-container {
+    position: fixed;
+    bottom: 0;
+    z-index: 2;
+    background-color: var(--white);
+    width: auto;
+    transform: translateX(-1.25rem);
+
+    .reply-view {
+      display: flex;
+      align-items: center;
+      justify-content: space-between;
+      padding: var(--size-1) var(--size-6);
+      background: var(--gray200);
+      color: var(--gray600);
+      font-size: var(--size-3);
+      font-weight: 400;
+      button svg {
+        vertical-align: bottom;
+      }
+    }
+
+    color: #747575;
+    .comment-view {
+      padding: 0 var(--size-5);
+      height: 4rem;
+      border-top: 1px solid var(--gray300);
+      display: flex;
+      align-items: center;
+      justify-content: center;
+      text-decoration-line: none;
+      input {
+        width: 16.6rem;
+      }
+      .add-comment {
+        display: inline;
+        border: none;
+        padding: var(--size-3) var(--size-5);
+        font-size: var(--size-4);
+        &::placeholder {
+          color: var(--gray500);
+        }
+        &:focus {
+          outline: none;
+        }
+        &:focus-visible {
+          outline: 1px solid transparent;
+        }
+      }
+      button {
+        width: 2.5rem;
+        color: var(--gray500);
+        &:hover {
+          color: var(--orange600);
+        }
+      }
+    }
+  }
+
+  .like-container {
+    margin: 0 0 0 1.875rem;
+
+    .like-title {
+      margin-bottom: var(--size-6);
+    }
+    div {
+      display: flex;
+      align-items: center;
+      justify-content: left;
+      .like-profile {
+        width: 3.5rem;
+        height: 3.5rem;
+        background: #eeeeee;
+        border-radius: 8px;
+        border: 1.3px solid #9e9e9e;
+        object-fit: cover;
+        margin-right: var(--size-5);
+      }
+      p {
+        font-weight: 500;
+        font-size: 0.875rem;
+        color: var(--gray1000);
+      }
+    }
+  }
+}
+
+.add-diary {
+  position: absolute;
+  bottom: calc(100% + var(--size-2));
+  left: calc(50% + 123px);
+  background-color: var(--orange600);
+  width: var(--size-12);
+  height: var(--size-12);
+  border-radius: 50%;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
