+++ src/Pages/DiaryNewPage/DiaryNewPage.module copy.scss
@@ -1,0 +1,61 @@
+.container {
+  width: 100%;
+  height: 100%;
+  overflow-y: scroll;
+  &::-webkit-scrollbar {
+    width: 16px;
+  }
+
+  &::-webkit-scrollbar-track {
+    background-color: transparent;
+  }
+
+  &::-webkit-scrollbar-thumb {
+    border: 4px solid transparent;
+    background: #606060;
+    background-clip: content-box;
+    border-radius: 8px;
+  }
+  margin: var(--size-1) var(--size-1) 0 var(--size-1);
+  border-top: 1px dashed var(--gray400);
+  border-left: 1px dashed var(--gray400);
+  border-radius: var(--size-1);
+  .read-container {
+    width: 100%;
+    display: flex;
+    justify-content: end;
+    padding-right: var(--size-4);
+    button {
+      font-weight: 400;
+      font-size: var(--size-3);
+      color: var(--gray500);
+
+      &:hover {
+        color: var(--orange600);
+      }
+    }
+  }
+  .header {
+    padding: 0 var(--size-5);
+    margin-top: var(--size-5);
+    h1 {
+      font-weight: 600;
+      font-size: 1.4375rem;
+      margin-bottom: var(--size-4);
+    }
+    p {
+      font-weight: 300;
+      color: var(--gray700);
+      font-size: var(--size-4);
+      margin-bottom: var(--size-5);
+    }
+  }
+  .content {
+    padding: 0 var(--size-5);
+    margin-top: var(--size-5);
+  }
+}
+
+.not-diary-container {
+  margin-top: 7rem;
+}
