+++ src/Pages/DiaryNewPage/DiaryNewPage.module.scss
@@ -1,61 +1,101 @@
-.container {
+.all-list {
+  display: flex;
+  flex-direction: column;
+  gap: 4.5rem;
+}
+
+.day-list-section {
+  display: flex;
+  flex-direction: column;
+  gap: 1.5rem;
+}
+
+.diary-list {
+  display: flex;
+  flex-direction: column;
+  gap: 1.5rem;
+  user-select: none;
+}
+
+.diary-item {
   width: 100%;
-  height: 100%;
-  overflow-y: scroll;
-  &::-webkit-scrollbar {
-    width: 16px;
+  border-radius: 0.5rem;
+  overflow: hidden;
+  box-shadow: 0 0 6px 0 rgba(0, 0, 0, 0.2);
+  cursor: pointer;
+
+  transition: transform 0.1s ease-in-out;
+
+  &:hover {
+    transform: scale(1.02);
+    z-index: 1;
   }
 
-  &::-webkit-scrollbar-track {
-    background-color: transparent;
-  }
-
-  &::-webkit-scrollbar-thumb {
-    border: 4px solid transparent;
-    background: #606060;
-    background-clip: content-box;
-    border-radius: 8px;
-  }
-  margin: var(--size-1) var(--size-1) 0 var(--size-1);
-  border-top: 1px dashed var(--gray400);
-  border-left: 1px dashed var(--gray400);
-  border-radius: var(--size-1);
-  .read-container {
-    width: 100%;
-    display: flex;
-    justify-content: end;
-    padding-right: var(--size-4);
-    button {
-      font-weight: 400;
-      font-size: var(--size-3);
-      color: var(--gray500);
-
-      &:hover {
-        color: var(--orange600);
-      }
-    }
-  }
-  .header {
-    padding: 0 var(--size-5);
-    margin-top: var(--size-5);
-    h1 {
-      font-weight: 600;
-      font-size: 1.4375rem;
-      margin-bottom: var(--size-4);
-    }
-    p {
-      font-weight: 300;
-      color: var(--gray700);
-      font-size: var(--size-4);
-      margin-bottom: var(--size-5);
-    }
-  }
-  .content {
-    padding: 0 var(--size-5);
-    margin-top: var(--size-5);
+  &:active {
+    transform: scale(0.98);
+    z-index: 1;
   }
 }
 
-.not-diary-container {
-  margin-top: 7rem;
+.diary-image-wrapper {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  width: 100%;
+  aspect-ratio: 5 / 4;
+  background-color: var(--white);
+
+  img {
+    width: 100%;
+    height: 100%;
+    object-fit: cover;
+  }
+}
+
+.row-group {
+  display: flex;
+  flex-direction: row;
+  gap: 0.5rem;
+}
+
+.col-group {
+  display: flex;
+  flex-direction: column;
+  gap: 0.5rem;
+}
+
+.size18 {
+  width: 1.125rem;
+  aspect-ratio: 1 / 1;
+}
+
+.with-icon {
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  justify-content: center;
+  gap: 0.25rem;
+
+  color: var(--gray700);
+}
+
+.relative-label {
+  color: var(--gray700);
+}
+
+.gray700 {
+  color: var(--gray700);
+}
+.no-diary-view {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  gap: 0.5rem;
+  padding-top: 3rem;
+}
+
+.size120 {
+  width: 7.5rem;
+  aspect-ratio: 1 / 1;
 }
