+++ src/Pages/MyBoardPage/MyBoardPage.module.scss
@@ -5,6 +5,11 @@
   width: 100%;
 }
 
+.header {
+  width: 100%;
+  height: fit-content;
+}
+
 .error-container {
   display: flex;
   flex-direction: column;
