+++ src/Styles/global.css
@@ -6,6 +6,7 @@
 
 div#root {
   display: flex;
+  flex-direction: column;
   min-height: 100dvh;
   max-height: 100dvh;
   margin: 0 auto;
